fix: resolve dash and weapon collisions after physics
